+++ Model/RouteGeneratorInterface.php
@@ -3,6 +3,7 @@
 namespace LoremIpsum\RouteGeneratorBundle\Model;
 
 use LoremIpsum\RouteGeneratorBundle\Exception\MissingRouteHandlerException;
+use Symfony\Component\Routing\Generator\UrlGeneratorInterface;
 
 interface RouteGeneratorInterface
 {
@@ -10,10 +11,11 @@ interface RouteGeneratorInterface
      * @param mixed       $value
      * @param string|null $view
      * @param array       $context
+     * @param int         $referenceType
      * @return mixed
      * @throws MissingRouteHandlerException
      */
-    public function generate($value, $view = null, $context = []);
+    public function generate($value, ?string $view = null, array $context = [], int $referenceType = UrlGeneratorInterface::ABSOLUTE_PATH);
 
     public function addRouteHandler(RouteHandlerInterface $handler);