+++ Model/RouteHandlerInterface.php
@@ -4,5 +4,12 @@ namespace LoremIpsum\RouteGeneratorBundle\Model;
 
 interface RouteHandlerInterface
 {
-    public function handle($value, $view = null, $context = []);
+    /**
+     * @param mixed       $value
+     * @param string|null $view
+     * @param array       $context
+     * @param int         $referenceType
+     * @return mixed
+     */
+    public function handle($value, ?string $view, array $context, int $referenceType);
 }