+++ Utils/RouteGenerator.php
@@ -5,6 +5,7 @@ namespace LoremIpsum\RouteGeneratorBundle\Utils;
 use LoremIpsum\RouteGeneratorBundle\Exception\MissingRouteHandlerException;
 use LoremIpsum\RouteGeneratorBundle\Model\RouteGeneratorInterface;
 use LoremIpsum\RouteGeneratorBundle\Model\RouteHandlerInterface;
+use Symfony\Component\Routing\Generator\UrlGeneratorInterface;
 use Symfony\Component\Routing\RouterInterface;
 
 class RouteGenerator implements RouteGeneratorInterface
@@ -24,10 +25,10 @@ class RouteGenerator implements RouteGeneratorInterface
         $this->router = $router;
     }
 
-    public function generate($value, $view = null, $context = [])
+    public function generate($value, ?string $view = null, array $context = [], int $referenceType = UrlGeneratorInterface::ABSOLUTE_PATH)
     {
         foreach ($this->handlers as $handler) {
-            $route = $handler->handle($value, $view, $context);
+            $route = $handler->handle($value, $view, $context, $referenceType);
             if ($route) {
                 return $route;
             }